requirements.txt
asgiref==3.7.2
astroid==2.3.3
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.3
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isort==4.3.21
joblib==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.1.1
nltk==3.7
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
pylint==2.4.4
python3-openid==3.2.0
pytube==9.6.0
pytube3==9.6.4
pytz==2023.3
regex==2022.6.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.14.0
sqlparse==0.4.4
tqdm==4.64.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.3
wrapt==1.11.2

